backend/requirements.txt
aniso8601==7.0.0
asgiref==3.4.1
autopep8==1.6.0
Django==3.2.11
djangorestframework==3.13.1
djangorestframework-simplejwt==5.0.0
graphene==2.1.9
graphene-django==2.15.0
graphql-core==2.3.2
graphql-relay==2.0.1
Pillow==9.0.0
promise==2.3
pyasn1==0.4.8
pycodestyle==2.8.0
pycryptodome==3.12.0
PyJWT==2.3.0
python-dotenv==0.19.2
pytz==2021.3
rsa==4.8
Rx==1.6.1
singledispatch==3.7.0
six==1.16.0
sqlparse==0.4.2
text-unidecode==1.3
toml==0.10.2
typing-extensions==4.0.1
